+++ src/common/js/constants.js
@@ -1,0 +1,5 @@
+/*
+ * pwix:bootbox/src/common/js/constants.js
+ */
+
+I18N = "pwix:bootbox:i18n:namespace";