+++ src/common/js/i18n.js
@@ -2,6 +2,8 @@
  * pwix:bootbox/src/common/js/i18n.js
  */
 
+import { pwixI18n } from 'meteor/pwix:i18n';
+
 pwixBootbox.i18n = {
     'en_US': {
         alert: {
@@ -32,3 +34,12 @@ pwixBootbox.i18n = {
 
 pwixBootbox.i18n.en = pwixBootbox.i18n.en_US;
 pwixBootbox.i18n.fr = pwixBootbox.i18n.fr_FR;
+
+pwixI18n.namespace( I18N, pwixBootbox.i18n );
+
+/**
+ * @returns {String} the i18n namespace of this package
+ */
+pwixBootbox.i18n.namespace = function(){
+    return I18N;
+};