+++ src/common/js/index.js
@@ -3,4 +3,6 @@
  */
 
 import './global.js';
+import './constants.js';
+//
 import './i18n.js';